정렬 메소드 수정

+++ src/main/java/lotto/Calculation.java
@@ -3,7 +3,7 @@ package lotto;
 public class Calculation {
     public static String calculate(int price) {
         //(총 당첨금 / 구매금액) * 100
-        float rate = totalAwards() / price * 100;
+        float rate = (float) totalAwards() / price * 100;
         return String.format("%.1f", rate);
     }
 
@@ -13,7 +13,7 @@ public class Calculation {
 
     public static int totalAwards() {
         int total = 0;
-        for(Ranks rank : Ranks.values()) {
+        for (Ranks rank : Ranks.values()) {
             total += doubles(rank);
         }
         return total;
